% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rk2libminer-package.R
\docType{package}
\name{rk2libminer-package}
\alias{rk2libminer}
\alias{rk2libminer-package}
\title{rk2libminer: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hi2rajendra/rk2libminer}
  \item Report bugs at \url{https://github.com/hi2rajendra/rk2libminer/issues}
}

}
\author{
\strong{Maintainer}: Rajendra K. Labala \email{rajendra.labala@gmail.com} (\href{https://orcid.org/0000-0003-0100-1868}{ORCID})

}
\keyword{internal}
